% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-defunct.R
\name{contact-defunct}
\alias{contact-defunct}
\title{Defunct functions in contact}
\description{
These functions have been removed from our package.
}
\details{
\itemize{
 \item \code{\link{contactTest}}: contactTest is defunct and was removed 
 in version 1.2.0. Please consider using another contact-comparison function
 instead (e.g., contactCompare_chisq, contactCompare_mantel, etc.)"
}
}
