% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactDur.all.R
\name{contactDur.all}
\alias{contactDur.all}
\title{Identify Interpersonal Contacts}
\usage{
contactDur.all(x, dist.threshold = 1, sec.threshold = 10,
  blocking = TRUE, blockUnit = "hours", blockLength = 1,
  equidistant.time = FALSE, parallel = TRUE, reportParameters = TRUE)
}
\arguments{
\item{x}{Output from the dist.all function. Can be either a data frame or list.}

\item{dist.threshold}{Description imminent}

\item{sec.threshold}{Description imminent}

\item{blocking}{Description imminent}

\item{blockUnit}{Description imminent}

\item{blockLength}{Description imminent}

\item{equidistant.time}{Description imminent}

\item{parallel}{Description imminent}

\item{reportParameters}{Description imminent}
}
\description{
This function uses the output from dist.all to determine when tracked individuals are in "contact" with one another. Individuals are said to be in a "contact" event (h) if they are observed within a given distance (<= dist.threshold) at a given timestep(i). Sec.threshold dictates the maximum amount of time a single, potential "contact" event should exist. For example, if sec.threshold=10, only "contacts" occurring within 10secs of one another will be regarded as a single "contact" event of duration sum(h). If in this case, a time difference between contacts was 11 seconds, the function will report two separate contact events.
}
\examples{
Examples imminent
}
\keyword{contact}
\keyword{data-processing}
