% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactTest.R
\name{contactTest}
\alias{contactTest}
\title{Determine if Observed Contacts are More or Less Frequent than in a Random Distribution}
\usage{
contactTest(emp.input, rand.input, dist.input, test = "chisq",
  indivComparison = TRUE, overallComparison = TRUE, blocking = TRUE,
  blockUnit = "mins", blockLength = 10, shuffle.type = 0,
  parallel = TRUE)
}
\arguments{
\item{emp.input}{List or data frame containing summarizeContact-function output refering to the empirical data.}

\item{rand.input}{List or data frame containing averaged summarizeContact-function output (i.e., argument "avg" was set to "TRUE" when the summarizeContacts function was run) refering to the randomized-path data.}

\item{dist.input}{List or data frame containing dist.all-/distToArea-function output refering to the empirical data.}

\item{test}{Describes the statistical test used to evaluate differences. Currently only takes the value "chisq," more tests will be added in later versions.}

\item{indivComparison}{Logical. If TRUE output will return a data frame showing individual-level comparsions. Defaults to TRUE.}

\item{overallComparison}{Logical. If TRUE output will return a data frame showing population-level comparisons. Defaults to TRUE.}

\item{blocking}{Description imminent}

\item{blockUnit}{Description imminent}

\item{blockLength}{Description imminent}

\item{shuffle.type}{Description imminent}

\item{parallel}{Description imminent}
}
\description{
This function is used to determine if tracked individuals in an empirical dataset had more or fewer contacts with other tracked individuals/specified locations than would be expected at random. The function works by comparing an empirically-based summarizeContacts function output (x) to the contactReport.avg generated from randomized data (y). Users can specify if they want to look at individual or overall (population-level) comparisons.
}
\details{
The default tested column (i.e., categorical data column from which data is drawn to be compared to randomized sets herein) is "id." This means that contacts involving each individual (defined by a unique "id") will be compared to randomized sets. Users may not use any data column for analysis other than "id." If users want to use another categorical data column in analyses rather than "id," we recommend re-processing data (starting from the dist.all/distToArea functions), while specifying this new data as an "id." For example, users may annotate an illness status column to the empirical input, wherein they describe if the tracked individual displayed gastrointestinal ("gastr"), respiratory ("respr"), both ("both"), illness symptoms, or were consistently healthy ("hel") over the course of the tracking period. Users could set this information as the "id," and carry it forward as such through the data-processing pipeline. Ultimately, they could determine if each of these disease states affected contact rates, relative to what would be expected at random.

Note: Currently, this function requires defined TSWs (temporal sampling windows; see the tempAggregate function) (i.e., contact durations must represent time periods that are equidistant).

Note: The current functionality is limited to comparisons using the X-squared test of independence. This works by comparing the number of TSWs in a given dataset/block between individuals/fixed locations and the number of TSWs during which individuals were observed, but were not in contact with specific individuals/places, for empirical (x) and randomized (y) datasets. The dist (z ; from dist.all or distToArea output) input is used here to determine how frequently each individual was observed in the empirical dataset/block of interest, allowing us to calculate the number of TSWs each individual was present but not involved in contacts. Note here that if X-squared expected values will be very small, approximations of p may not be right (and in fact, all estimates will be poor). It may be best to weight these tests differently. To address this, I've added the "warning" column to the output which notifies users when the chi-sq function reported that results may be inaccurate.

Note: if blocking == TRUE, blockLength, and blockUnit should be consistent with previous function inputs.
}
\examples{
Examples imminent
}
\keyword{data-processing}
\keyword{location}
\keyword{point}
\keyword{smoothing}
