% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFake.R
\name{dateFake}
\alias{dateFake}
\title{Create Fake Date Information}
\usage{
dateFake(timestamp, dateFormat = "mdy", startYear = 2000)
}
\arguments{
\item{timestamp}{Description imminent}

\item{dateFormat}{Description imminent}

\item{startYear}{Description imminent}
}
\description{
This function assigns fake date information, beginning 01/01/startYear, to each of the timestamps. Users can control what format the output vector is in by changing the dateFormat argument (format: "mdy" =  month-day-year, "ymd" =  year-month-day, "dmy" =  day-month-year, or "ydm" =  year-day-month).
}
\details{
Note that the timestamp argument should be a vector of all relevant timepoints. Additionally, timepoints should by in hms ("hour, minute, second") format.
}
\examples{
Examples imminent
}
\keyword{data-processing}
