% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.append.R
\name{datetime.append}
\alias{datetime.append}
\title{Append Date-Time Information to a Dataset}
\usage{
datetime.append(x, date = NULL, time = NULL, dateTime = NULL,
  dateFormat = "mdy", dateFake = FALSE, startYear = 2000,
  tz.in = "UTC", tz.out = NULL, month = FALSE, day = FALSE,
  year = FALSE, hour = FALSE, minute = FALSE, second = FALSE,
  daySecond = FALSE, dayID = FALSE, startID = 1,
  totalSecond = FALSE)
}
\arguments{
\item{x}{List or data frame to which new information will be appended.}

\item{date}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "date" exists in x, or that the dateTime argument != NULL. Defaults to NULL.}

\item{time}{Description imminent}

\item{dateTime}{Description imminent}

\item{dateFormat}{Description imminent}

\item{dateFake}{Logical. If dateFake == TRUE, the function will assign fake date information, beginning 01/01/startYear, to each of the timestamps. Defaults to FALSE.}

\item{startYear}{Description imminent}

\item{tz.in}{Character. Identifies the timezone associated with the time/dateTime argument input. Defaults to "UTC." Timezone names often take the form "Country/City." See the listing of timezones at: http://en.wikipedia.org/wiki/List_of_tz_database_time_zones.}

\item{tz.out}{Character. Identifies the timezone that the output dateTime information will be converted to. If NULL, tz.out will be identical to tz.in. Defaults to NULL. Timezone names often take the form "Country/City." See the listing of timezones at: http://en.wikipedia.org/wiki/List_of_tz_database_time_zones.}

\item{month}{Description imminent}

\item{day}{Description imminent}

\item{year}{Description imminent}

\item{hour}{Description imminent}

\item{minute}{Description imminent}

\item{second}{Description imminent}

\item{daySecond}{Description imminent}

\item{dayID}{Description imminent}

\item{startID}{Description imminent}

\item{totalSecond}{Description imminent}

\item{changeTimezone}{Description imminent}
}
\description{
This function appends date-time information to a dataset in POSIXct date_time format. It also uses functions from the lubridate package and minor calculations to parse out month, day, hour, minute, second, daySecond (the sequentially ordered second of a day), and totalSecond (sequentially ordered second over the course of the study period) of observations in a given dataset with date (format: "mdy" =  month/day/year, "ymd" =  year/month/day, "dmy" =  day/month/year, or "ydm" =  year/day/month (note: no preceding zeroes should be included before numbers <10)) and time (format: hour:minute:second (note:preceding zeroes must be included before numbers < 10, ex. 00:00:01)) information, appends this metadata to the dataset,and can assign each day a unique ID.
}
\examples{
Examples imminent
}
\keyword{date}
\keyword{date-time}
\keyword{time}
