% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2Area_df.R
\name{dist2Area_df}
\alias{dist2Area_df}
\title{Calculate Distances Between Individuals and Fixed Points/Polygons}
\usage{
dist2Area_df(x = NULL, y = NULL, x.id = NULL, y.id = NULL,
  dateTime = NULL, point.x = NULL, point.y = NULL, poly.xy = NULL,
  parallel = TRUE, dataType = "Point", lonlat = FALSE,
  numVertices = 4)
}
\arguments{
\item{x}{Description imminent}

\item{y}{Description imminent}

\item{x.id}{Description imminent}

\item{y.id}{Description imminent}

\item{dateTime}{Description imminent}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{poly.xy}{Description imminent}

\item{parallel}{Description imminent}

\item{dataType}{Description imminent}

\item{lonlat}{Description imminent}

\item{numVertices}{Description imminent}
}
\description{
Calculate distances (either planar or great circle - see dist.all) between each individual, reported in x, and a fixed point(s)/polygon(s), reported in y, at each timestep.
}
\details{
Polygon coordinates (in both x and y inputs) must be arranged in the format of those in referencePointToPolygon outputs (i.e., col1 = point1.x, col2 = point1.y, col3 =point2.x, col4 = point2.y, etc., with points listed in a clockwise (or counter-clockwise) order).
}
\examples{
Examples imminent
}
\keyword{GRC}
\keyword{data-processing}
\keyword{location}
\keyword{planar}
\keyword{point}
\keyword{polygon}
