% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2Area_sf.R
\name{dist2Area_sf}
\alias{dist2Area_sf}
\title{Calculate Distances Between Individuals and Fixed Points/Polygons}
\usage{
dist2Area_sf(x, y, x.id = NULL, y.id = NULL, dateTime = NULL,
  parallel = TRUE)
}
\arguments{
\item{x}{Description imminent}

\item{y}{Description imminent}

\item{x.id}{Description imminent}

\item{y.id}{Description imminent}

\item{dateTime}{Description imminent}

\item{parallel}{Description imminent}
}
\description{
Calculate distances between each individual, reported in x, and a fixed point(s)/polygon(s), reported in y, at each timestep. In this case, we suggest removing duplicate entries with the dup function and converting the input to a SpatialPointsDataFrame prior to running the function.
}
\details{
Inputs x and y must be spatial objects of class sf (i.e., shapefile), SpatialPointsDataFrame, or SpatialPolygonsDataFrame.
If inputs are SpatialMultiPointsDataFrame, which have one-to-many cardinality, you will receive this error: "Error in .pointsToMatrix(p) : points should be vectors of length 2, matrices with 2 columns, or inheriting from a SpatialPoints* object."
}
\examples{
Examples imminent
}
\keyword{data-processing}
\keyword{location}
\keyword{point}
\keyword{polygon}
\keyword{spatial}
