% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.calc.R
\name{dt.calc}
\alias{dt.calc}
\title{Calculate Time Difference Between Relocations}
\usage{
dt.calc(x = NULL, id = NULL, dateTime1 = NULL, dateTime2 = NULL,
  timeUnits = "secs", parallel = TRUE, timeStepRelation = 1)
}
\arguments{
\item{x}{Description imminent}

\item{id}{Description imminent}

\item{dateTime1}{Description imminent}

\item{dateTime2}{Description imminent}

\item{timeUnits}{Description imminent}

\item{parallel}{Description imminent}

\item{timeStepRelation}{Description imminent}
}
\description{
This function calculates the time difference between relocation events, accounting for individuals' ids. This function has the capability to calculate the differences between sequential timepoints related to two different features (e.g., contactStartTime and contactEndTime) if both dateTime1 and dateTime2 are defined, or just sequential timepoints from a single vector (e.g., contactStartTime) if only dateTime1 is defined.
}
\examples{
Examples imminent
}
\keyword{contact}
\keyword{data-processing}
