% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup.R
\name{dup}
\alias{dup}
\title{Identify and Remove Duplicated Data Points}
\usage{
dup(x, id = NULL, point.x = NULL, point.y = NULL, dateTime = NULL,
  avg = TRUE, parallel = TRUE, filterOutput = TRUE)
}
\arguments{
\item{x}{List or data frame that will be filtered.}

\item{id}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "id" exists in x. Defaults to NULL.}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{dateTime}{Description imminent}

\item{avg}{Description imminent}

\item{parallel}{Description imminent}

\item{filterOutput}{Description imminent}
}
\description{
dup (a.k.a. Multiple instance filter) identifies and removes timepoints when tracked individuals were observed in >1 place concurrently. If avg == TRUE, duplicates are replaced by a single row describing an individuals' average location (e.g., planar xy coordinates) during the duplicated time point. If avg == FALSE, all duplicated timepoints will be removed, as there is no way for the function to determine which instance among the duplicates should stay. If users are not actually interested in filtering datasets, but rather, determining what observations should be filtered, they may set filterOutput == FALSE. By doing so, this function will append a "duplicated" column to the dataset, which reports values that describe if any timepoints in a given individual's path are duplicated. Values are: 0: timepoint is not duplicated, 1: timepoint is duplicated.
}
\details{
If users want to remove specific duplicated observations, we suggest setting filterOutput == FALSE, reviewing what duplicated timepoints exist in individuals' paths, and manually removing observations of interest.
}
\examples{
Examples imminent
}
\keyword{duplicates}
\keyword{filter}
