% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDistThresh.R
\name{findDistThresh}
\alias{findDistThresh}
\title{Identify Distance Threshold for Contact}
\usage{
findDistThresh(n1 = 1000, n2 = 1000, acc.Dist1 = 0.5,
  acc.Dist2 = NULL, pWithin1 = 90, pWithin2 = NULL, delta = 0.666)
}
\arguments{
\item{n1}{Number of points used in the expected-distance distribution(s). Defaults to 1000.}

\item{n2}{Number of expected-distance distribution iterations to be averaaged. Defaults to 1000.}

\item{acc.Dist1}{Accuracy distance for point 1.}

\item{acc.Dist2}{Accuracy distance for point 2. If == NULL, defaults to acc.Dist1 value.}

\item{pWithin1}{Percent of data points within acc.Dist of true locations for point 1.}

\item{pWithin2}{Percent of data points within acc.Dist of true locations for point 2. If == NULL, defaults to pWithin1 value.}

\item{delta}{Length representing biological mechanism for contact.}
}
\description{
Identify Distance Threshold for Contact
}
\examples{
Examples imminent
}
\keyword{contact}
\keyword{location}
\keyword{point}
