% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{mps}
\alias{mps}
\title{Identify and Remove Data Points Based on Observed Movement Speed}
\usage{
mps(x, id = NULL, point.x = NULL, point.y = NULL, dateTime = NULL,
  mpsThreshold = 10, lonlat = FALSE, parallel = TRUE,
  filterOutput = TRUE)
}
\arguments{
\item{x}{List or data frame that will be filtered.}

\item{id}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "id" exists in x. Defaults to NULL.}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{dateTime}{Description imminent}

\item{mpsThreshold}{Description imminent}

\item{lonlat}{Description imminent}

\item{parallel}{Description imminent}

\item{filterOutput}{Description imminent}
}
\description{
mps (a.k.a. Movement-per-Second Filter) identifies and removes timepoints when tracked individuals were observed moving faster than a set distance threshold (representing either the great-circle distance between two points a planar distance metric, depending on whether or not lonlat == TRUE or FALSE, respectively) per second. (i.e., if it is impossible/highly unlikely that individuals moved faster than a given speed (mps), we can assume that any instances when they were observed doing so were the result of erroneous reporting, and should be removed). When running the mps filter, users have the option of setting lonlat == TRUE (by default lonlat == FALSE). lonlat is a logical argument that tells the function to calculate the distance between points on the WGS ellipsoid (if lonlat == TRUE), or on a plane (lonlat == FALSE) (see raster::pointDistance). If lonlat == TRUE, coordinates should be in degrees. Otherwise, coordinates should represent planar ('Euclidean') space (e.g. units of meters).
}
\details{
If users are not actually interested in filtering datasets, but rather, determining what observations should be filtered, they may set filterOutput == FALSE. By doing so, this function will append up an "mps" column to the dataset, which reports the avg distance per second individuals moved to get from observation i-1 to observation i.
}
\examples{
Examples imminent
}
\keyword{filter}
\keyword{movement}
\keyword{speed}
