% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeFeature.R
\name{randomizeFeature}
\alias{randomizeFeature}
\title{Randomize or Pseudorandomize Categorical Variables Associated With Individuals}
\usage{
randomizeFeature(x, feature = NULL, shuffle = FALSE,
  maintainDistr = TRUE, numRandomizations = 10, parallel = TRUE)
}
\arguments{
\item{x}{Description imminent}

\item{feature}{Description imminent}

\item{shuffle}{Description imminent}

\item{maintainDistr}{Description imminent}

\item{numRandomizations}{Description imminent}
}
\description{
This function randomizes the values in a given column (or set of columns (i.e., c(colname(x)[1], colname(x)[2]))), identified by "feature," in a dataset(x). If maintainDistr == TRUE, the number of each unique value in the column will be maintained in the function output. Otherwise, the function will draw on the initial distribution to assign randomized values, but the specific number of each unique value may not be maintained. If shuffle == TRUE, unique values will be replaced with another, random unique value from the same distribution with 100% certainty. For example if the values in a "dose" column c(0mg,100mg,300mg) were shuffled, one possible outcome would be: x$dose.shuff[which(x$dose == "0mg")] <- 300mg, x$dose.shuff[which(x$dose == "100mg")] <- 0mg, and x$dose.shuff[which(x$dose == "300mg")] <- 100mg.
}
\details{
Note: the only subfunction utilized in this function (for which parallel is relevant) is used ONLY if shuffle == TRUE.
}
\examples{
Examples imminent
}
\keyword{analysis}
\keyword{randomize}
