% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizePaths.R
\name{randomizePaths}
\alias{randomizePaths}
\title{Randomize or Pseudorandomize Individuals' Relocation Events}
\usage{
randomizePaths(x = NULL, id = NULL, dateTime = NULL,
  point.x = NULL, point.y = NULL, poly.xy = NULL, parallel = TRUE,
  dataType = "Point", numVertices = 4, blocking = TRUE,
  blockUnit = "mins", blockLength = 10, shuffle.type = 0,
  shuffleUnit = "Days", indivPaths = TRUE, numRandomizations = 10)
}
\arguments{
\item{x}{Description imminent}

\item{id}{Description imminent}

\item{dateTime}{Description imminent}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{poly.xy}{Description imminent}

\item{parallel}{Description imminent}

\item{dataType}{Description imminent}

\item{numVertices}{Description imminent}

\item{blocking}{Description imminent}

\item{blockUnit}{Description imminent}

\item{blockLength}{Description imminent}

\item{shuffle.type}{Description imminent}

\item{shuffleUnit}{Description imminent}

\item{indivPaths}{Description imminent}

\item{numRandomizations}{Description imminent}
}
\description{
Randomizes individuals' spatial locations. Randomized datasets can later be compared to empirical ones to determine if individuals' space use differ from what would be expected at random (using the contactTest function).
}
\details{
Paths can be randomized, or pseudorandomized differently according to what logical arguments are set to TRUE. 
1.) If indivPaths == TRUE, paths will be randomized with no location switching between ids (e.g., randomized xy locations for individual 1 will be generated by sampling only from individual 1's location distribution). If indivPaths == FALSE, paths will be randomized with potential location switching between ids (e.g., randomized xy locations for individual 1 will be generated by sampling from the entire dataset's location distribution)
2.) If blocking == FALSE, paths will be randomized by sampling from observed timepoints. No timepoints will be represented more than once in the randomized set. If blocking == TRUE, prior to randomization, timepoints will be categorized into a series of temporal blocks of blockLength-blockUnit length (e.g., 10 mins). After generating blocks, the spatial-location randomization methodology will follow shuffle.type.
3.) If shuffle.type == 0, within-block timepoints will be randomized by sampling from observed timepoints only within the relevant block (e.g., points in block 1 may be redistributed). Block order in the dataset does not change, and no timepoints will be represented more than once in the randomized set. If shuffle.type == 1, blocks are shuffled around, but points within blocks are not redistributed (e.g., block 1 <- block 5, block 3 <- block 2, block 5 <- block 4). If shuffle.type == 2, blocks are shuffled, but their relative temporal location in the shuffleUnit is maintained. For example, there are 144 10-min blocks in a 24-hour day. Say our datset contains 3 days worth of data (i.e., 432 blocks). During the randomization, because blocks maintain their relative locations in shuffleUnits (e.g., Days), block 1 in the random set will be determined by sampling from a distribution of blocks 1,145,and 289, which each representing the first block of a given shuffleUnit (e.g., Day 1, Day 2, Day 3). All blocks in the randomized set are decided in this fashion (e.g., block 2 of the randomized set is identified by sampling from a distribution of blocks 2, 146, and 290). Therefore, blocklength-unit may never exceed 1 shuffleUnit (e.g., 25-hour blocks will cannot be shuffled using shuffleUnit == "Days," but 1-24-hour blocks work just fine. Points within blocks are not redistributed.This particular shuffle.type (i.e., 2) is based off of the methodology described by Spiegel et al. 2016 (https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.12553).
}
\examples{
Examples imminent
}
\keyword{contact}
\keyword{data-processing}
