% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencePointToPolygon.R
\name{referencePointToPolygon}
\alias{referencePointToPolygon}
\title{Create a Rectangular Polygon Using Planar XY Coordinates}
\usage{
referencePointToPolygon(x = NULL, id = NULL, dateTime = NULL,
  point.x = NULL, point.y = NULL, StartLocation = "UL",
  UpDownRepositionLen = 1, LeftRightRepositionLen = 1,
  CenterPoint = FALSE, MidPoints = FALSE, StandingThreshold = 0,
  parallel = TRUE)
}
\arguments{
\item{x}{Description imminent}

\item{id}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "id" exists in x. Defaults to NULL.}

\item{dateTime}{Description imminent}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{StartLocation}{Description imminent}

\item{UpDownRepositionLen}{Description imminent}

\item{LeftRightRepositionLen}{Description imminent}

\item{CenterPoint}{Description imminent}

\item{MidPoints}{Description imminent}

\item{StandingThreshold}{Description imminent}

\item{parallel}{Description imminent}
}
\description{
This function creates a square/rectangular polygon from a single reference point by translating its location multiple times using the same method used in repositionReferencePoint. For example, even though calves in our study were only equiped with RFID tags on their left ear. With this function, we can create polygons that account for the total space used by each individual at each time step.This function is different from similar point-to-polygon functions for two reasons:
1.) It does not assume points lie within the center of the polygon. Rather, the reference point must be a corner of the polygon (Note: "UL" denotes that reference point lies on the upper-left corner of the polygon, "UR" denotes that reference point lies on the upper-right corner of the polygon,"DL" denotes that reference point lies on the down-left corner of the polygon, "DR" denotes that reference point lies on the down-left corner of the polygon). Note that if you want the reference point to be at the center of the polygon, you can first translate the reference point to a central location on tracked individuals using repositionReferencePoint.
2.) Polygon angles/directionality are based on observed movements of tracked individuals.
}
\details{
Currently, this function only supports input data with coordinates representing planar ('Euclidean') space (e.g. units of meters).

In the output, point1.x and point1.y represent the xy coordinates from the input file. Point2-n coordinates move in a clockwise direction from point1. For example: if point1 is located on the upper left ("UL") corner of the polygon, point2 would be on the upper right corner, point3 on the bottom right, and point 4 on the bottom left.

Because this function needs information (dist, dx, dy) from 2 points on an individual's path to work, at least the first point in each individual's path will be removed (the function will report NAs for adjusted locations). Also note that if the distance between an individual's first point in their path and the second one is 0, the function will also report NAs for the second point's adjusted coordinates. The first non-NA values will only be reported for the instance where dist > 0.

In the output, if input was previously processed using tempAggregate at a Day or Hour resolutionLevel, dt > secondAgg indicates that tracked individuals were missing in the original dataset on a given day or hour, respectively. In this case, the assumption that individuals are facing a given direction because they moved from the previous timepoint may not be accurate. Consider removing these rows (rows following one with dt > secondAgg; remember that dt indicates the time between recording xy coordinates in row i to row i + 1) from your dataset.
}
\examples{
Examples imminent
}
\keyword{data-processing}
\keyword{location}
\keyword{planar}
\keyword{point}
\keyword{polygon}
