% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeContacts.R
\name{summarizeContacts}
\alias{summarizeContacts}
\title{Display Contact Events}
\usage{
summarizeContacts(x, importBlocks = FALSE, avg = FALSE)
}
\arguments{
\item{x}{Output from the contactDur.all or contactDur.area functions. Can be either a data frame or list.}

\item{importBlocks}{Description imminent}

\item{avg}{Description imminent}
}
\description{
This function takes the output from contactDur.all or contactDur.area and reports the number of durations when tracked individuals are in "contact" with one another (contactDur.all) or with specified fixed points/polygons (contactDur.area).
}
\details{
If x is a list, and avg == TRUE, this function will produce an extra data frame containing the mean column values for each id (per block if importBlocks == TRUE).
}
\examples{
Examples imminent
}
\keyword{contact}
\keyword{data-processing}
