% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempAggregate.R
\name{tempAggregate}
\alias{tempAggregate}
\title{Smooth Location Data Points Over Time}
\usage{
tempAggregate(x = NULL, id = NULL, point.x = NULL, point.y = NULL,
  dateTime = NULL, secondAgg = 10, extrapolate.left = FALSE,
  extrapolate.right = FALSE, resolutionLevel = "Full",
  parallel = TRUE, na.rm = TRUE, smooth.type = 1)
}
\arguments{
\item{x}{List or data frame that will be filtered.}

\item{id}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "id" exists in x. Defaults to NULL.}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{dateTime}{Description imminent}

\item{secondAgg}{Description imminent}

\item{extrapolate.left}{Description imminent}

\item{extrapolate.right}{Description imminent}

\item{resolutionLevel}{Description imminent}

\item{parallel}{Description imminent}

\item{na.rm}{Description imminent}

\item{smooth.type}{Description imminent}
}
\description{
Aggregate location data by secondAgg seconds over the course of each day represented in the dataset. The function smooths xy data forwards (smooth.type == 1) or backwards (smooth.type == 2) according to a data-point-averaging loess smoothing methodology.As part of the smoothing process, tempAggregate fills in any missing values (either due to a lack of data transmission or faulty prior interpolation). We recognize that this procedure is not sensitive to individual presence at given timesteps (e.g., some individuals may be missing on certain days, hours, etc., and therefore may produce inaccurate location aggregates if days/hours exist where individuals are not present in the dataset (e.g., they were purposefully removed, or moved outside of the monitoring area)). To increase accuracy, package users may specify a resolutionLevel ("full" or "reduced") to process individuals' locations at different resolutions. If resolution == "reduced", if no locations of individuals exist over any secondAgg time block, NAs will be produced for the time blocks of interest.
}
\examples{
Examples imminent
}
\keyword{data-processing}
\keyword{location}
\keyword{point}
\keyword{smoothing}
