% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeBlock.append.R
\name{timeBlock.append}
\alias{timeBlock.append}
\title{Append TimeBlock Information to a Data Frame}
\usage{
timeBlock.append(x = NULL, dateTime = NULL, blockLength = 10,
  blockUnit = "mins")
}
\arguments{
\item{x}{Description imminent}

\item{dateTime}{Description imminent}

\item{blockLength}{Description imminent}

\item{blockUnit}{Description imminent}
}
\description{
Appends "block," "block.start," "block.end," and "numBlocks" columns to an input data frame (x) with a dateTime (see dateTime.append) column. This allows users to "block" data into blockLength-blockUnit-long (e.g., 10-min-long) temporal blocks. If x == NULL, the function output will be a data frame with "dateTime" and block-related columns.
}
\examples{
Examples imminent
}
\keyword{data-processing}
