% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositionReferencePoint.R
\name{repositionReferencePoint}
\alias{repositionReferencePoint}
\title{Move Data Point a Specified Distance}
\usage{
repositionReferencePoint(x = NULL, id = NULL, dateTime = NULL,
  point.x = NULL, point.y = NULL, RepositionDir = "Up",
  UpDownRepositionLen = 1, LeftRightRepositionLen = 1,
  StandingThreshold = 0, parallel = TRUE)
}
\arguments{
\item{x}{Description imminent}

\item{id}{Vector of length(nrow(data.frame(x))) or singular character data, detailng the relevant colname in x, that denotes what date information will be used. If argument == NULL, datetime.append assumes a column withe colname "id" exists in x. Defaults to NULL.}

\item{dateTime}{Description imminent}

\item{point.x}{Description imminent}

\item{point.y}{Description imminent}

\item{parallel}{Description imminent}

\item{secondAgg}{Description imminent}

\item{extrapolate.left}{Description imminent}

\item{extrapolate.right}{Description imminent}

\item{resolutionLevel}{Description imminent}

\item{na.rm}{Description imminent}

\item{smooth.type}{Description imminent}
}
\description{
Translates locations of a single rfid tag/gps transmitter to a different location on the equipped animal (i.e., move the coordinate up, down, left, right, or a combination thereof (i.e., up & right, up & left, down & right, down & left)) . For example, calves in our study (the test dataset) were equiped with RFID tags on their left ear. With this code, we can move this reference point somewhere else on the body of each individual. This might be done for a number of reasons, but is very useful for use in the referencePointToPolygon function later on (for delineating polygons representing entire individuals). Currently, this function only supports input data with coordinates representing planar ('Euclidean') space (e.g. units of meters).
}
\details{
In this function, if the distance individuals moved was less than/equal to the noted StandingThreshold, individuals are said to be "standing," and their position will not change relative to their previous one. (i.e., you assume that any observed movement less than StandingThreshold was due to errors or miniscule bodily movements (e.g., head shaking) that are not indicative of actual movement.)

Because this function needs information (dist, dx, dy) from 2 points on an individual's path to work, at least the first point in each individual's path will be removed (the function will report NAs for adjusted locations). Also note that if the distance between an individual's first point in their path and the second one is 0, the function will also report NAs for the second point's adjusted coordinates. The first non-NA values will only be reported for the instance where dist > 0.

In the output, if input was previously processed using tempAggregate at a Day or Hour resolutionLevel, dt > secondAgg indicates that tracked individuals were missing in the original dataset on a given day or hour, respectively. In this case, the assumption that individuals are facing a given direction because they moved from the previous timepoint may not be accurate. Consider removing these rows (rows following one with dt > secondAgg; remember that dt indicates the time between recording xy coordinates in row i to row i + 1) from your dataset.
}
\examples{
Examples imminent
}
\keyword{data-processing}
\keyword{location}
\keyword{planar}
\keyword{point}
